requirements.txt
selenium
robotframework>=3.1.2
robotframework-faker>=4.3.0
robotframework-seleniumlibrary>=4.1.0
robotframework-selenium2library>=3.0.0
robotframework-pageobjectlibrary>=1.0.2

coloredlogs~=15.0
features~=0.5.12